Refac

+++ pacman-core/src/main/java/de/amr/pacmanfx/Globals.java
@@ -68,7 +68,7 @@ public interface Globals {
      * @param numTiles number of tiles
      * @return number of pixels corresponding to given number of tiles
      */
-    static float TS(double numTiles) { return (float) (numTiles * TS); }
+    static float TS(float numTiles) { return numTiles * TS; }
 
     /**
      * Arcade maps have a size of 28 cols and 36 rows (including the empty rows over and under the maze).
@@ -76,8 +76,8 @@ public interface Globals {
     Vector2i ARCADE_MAP_SIZE_IN_TILES = new Vector2i(28, 36);
 
     /**
-     * Arcade maps have a size of 28x36 tiles (including the empty rows over and under the maze). The tile size is 8px
-     * which gives a map size of 224x288px.
+     * Arcade maps have a size of 28x36 tiles (28 cols, 36 rows, including the empty rows over and under the maze).
+     * The tile size is 8px which gives a map size of 224x288px.
      */
     Vector2i ARCADE_MAP_SIZE_IN_PIXELS = new Vector2i(224, 288); // 28x36 tiles
